fix(s3): Collapse config into a single configuration class
Since we ComponentScan everything under the sun, using `@Import` on configuration
classes is totally useless, so any attempt to break up configuration classes is
impossible.
